// sourcecode/BTrak.Packages/snova-hrmangement/module/module-components/src/lib/hrmanagment/models/employee-membership-details-model.ts
export class EmployeeMembershipDetailsModel{
    employeeMembershipId: string;
    employeeId: string;
    membershipId: string;
    subscriptionId: string;
    subscriptionAmount: number;
    currencyId: string;
    commenceDate: Date;
    renewalDate: Date;
    isArchived: boolean;
    timeStamp: any;
    nameOfTheCertificate: string;
    issueCertifyingAuthority: string;

    firstName: string;
    surName: string;
    userName: string;
    email: string;
    membership: string;
    subscriptionPaidByName: string;
    currencyName: string;
    totalCount: number;
}